Migrated to SDK2 and implemented homealarm device class.
